requirements.txt
Django==4.2.2
djangorestframework==3.14.0
APScheduler==3.10.1
